Add new library: multibuffer
Multibuffer is a thread-safe single-writer multiple-reader buffer.  It provides
an io.Writer which writes to an internal buffer, and multiple readers that can
read said data.

While nothing (besides tests) uses this now, the intended purpose is for use in
log tailing.  The job server will exec a process with a multibuffer attached to
its standard out and error.  When a client requests logs via RPC, they'll get a
reader and stream out logs.

This only implements the core synchronization primitive needed for log tailing;
some additional features will be needed but they can be built on top of this.
Being a standard io.Writer and io.Reader means it can be composed with other
parts of the Go ecosystem easily.